requirements.txt
PyQt6==6.6.1
openai-whisper==20231117
googletrans==4.0.0rc1
ffmpeg-python==0.2.0
torch>=2.0.0
numpy>=1.24.0
